[PM-19371] Blank Screen on Browser Extension Popup After Approving TDE Request on Popout (#14211)
* feat: clear cached history after successfull login

* Revert "feat: clear cached history after successfull login"

This reverts commit 4ede7f90564135da97c754b6b90b26a483cb6b98.

* feat: only restore router cache once per popup session

The purpose of the router cache is to restore the last visited route
during the startup of the popup, but without any explicit check
the cache would try to restore the route an inifinite number of times.
Because the router cache is never used to restore a route any other time
it is safe to assume that we only want to run the restore function once.