[PM-17407] [CL-665] Remove jQuery and Popper.js (#14621)
Now that the last usages of ModalService is removed from the web portions we can finally remove jQuery and Popper.js. This PR also removes bootstrap js imports since it would drag in jQuery as a peer dependency.

Note: Both dependencies still exists in the lockfile as they are peer dependencies of boostrap.